requirements.txt
streamlit>=1.32.2
python-dotenv>=1.0.1
requests>=2.31.0
openai-agents
pydantic
typing-extensions
fastapi
uvicorn
python-multipart
